# Backend-services/events-management/swagger.yaml
swagger: '2.0'

# Document metaData

info:
  version : "1.0.0"
  title: Events
# URL Data  
host: localhost
basePath: /events
tags: 
 - name: EVENTS
schemes: 
  - http
# End Points
paths:
  # get Authentication
  /eventid/{eventId}:
    get:
      tags:
       - Dashboard
      description: get event by id
   
      
      produces:
        - application/json
      parameters:
        - name: eventId
          in: path 
          required: true
          type: string

           
      responses: 
          200:
            description : fetch
            schema:
              $ref: '#/definitions/Event'
              
          404:
            description: No DataFound 
            
  /all:
    get:
      tags:
       - Dashboard
      description: get all events
   
      
      produces:
        - application/json

           
      responses: 
          200:
            description : fetch
            schema:
              $ref: '#/definitions/Event'
              
          404:
            description: No DataFound
  
  /summary/all:
    get:
      tags:
       - Event Summary
      description: get the summary of all events
   
      
      produces:
        - application/json

           
      responses: 
          200:
            description : fetch
            schema:
              $ref: '#/definitions/EventSummary'
              
          404:
            description: No DataFound 
  
  /summaryId/{summaryId}:
    get:
      tags:
       - Event Summary
      description: get event summary by id
   
      
      produces:
        - application/json
      parameters:
        - name: summaryId
          in: path 
          required: true
          type: integer

           
      responses: 
          200:
            description : fetch
            schema:
              $ref: '#/definitions/EventSummary'
              
          404:
            description: No DataFound             
          
            
definitions:

    
    Event:
      type: object
      properties:
        eventId:
          type: string
        baseLocation:
          type: string
        beneficiaryName:
          type: string
        councilName:
          type: string
        eventName:
          type: string
        eventDescription:
          type: string
        eventDate:
          type: string
          format: date
        employeeId:
          type: number
        employeeName:
          type: string
        volunteerHours:
          type: integer
        travelHours:
          type: integer
        livesImpacted:
          type: integer
        businessUnit:
          type: string
        status:
          type: string
        iiepCategory:
          type: string
        summaryId:
          type: integer
          
    EventSummary:
      type: object
      properties:
        summaryId:
          type: string
        eventId:
          type: string
        month:
          type: string
        baseLocation:
          type: string
        beneficiaryName:
          type: string
        venueAddress:
          type: string
        councilName:
          type: string
        project:
          type: string
        category:
          type: string
        eventName:
          type: string
        eventDescription:
          type: string
        eventDate:
          type: string
          format: date
        totalNoOfVolunteers:
          type: integer
        totalVolunteerHours:
          type: integer
        totalTravelHours:
          type: integer
        OverallVolunteerHours:
          type: integer
        livesImpacted:
          type: integer
        activityType:
          type: integer
        status:
          type: string
        pocId:
          type: number
        pocName:
          type: string
        pocContactnumber:
          type: number